v1.9.6

+++ src/main/java/com/inolia_zaicek/more_mod_tetra/Effect/MMT/Titan/FirstLightHealsTheWorld.java
@@ -6,6 +6,7 @@ import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.event.entity.living.LivingHealEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
+import com.inolia_zaicek.more_mod_tetra.Event.Post.EffectLevelEvent;
 import se.mickelus.tetra.blocks.workbench.gui.WorkbenchStatsGui;
 import se.mickelus.tetra.gui.stats.StatsHelper;
 import se.mickelus.tetra.gui.stats.bar.GuiStatBar;
@@ -36,7 +37,7 @@ public class FirstLightHealsTheWorld {
             LivingEntity player = event.getEntity();
             int effectLevel = MMTEffectHelper.getInstance().getMainOffHandMaxEffectLevel(player,firstLightHealsTheWorldEffect);
             if (effectLevel > 0) {
-                float number=1+(effectLevel/100 );
+                float number=1+((float) effectLevel /100 );
                 float heal = event.getAmount();
                 event.setAmount(heal*number);
             }
